Port workflows to Apple Silicon
Use macos-14 image.

Remove SDK download. We build now as-is on GitHub.

Do not interfere with /usr/local inside GitHub CI anymore. The image is
nearly clean in that regard.

Turn the check "Is /usr/local clean?" into a warning by default. This
warning will always appear in GitHub CI.